fix data, methods, and func's

+++ controller/contacts.js
@@ -1,12 +1,19 @@
 const { HttpError, ctrlWrapper } = require('../helpers')
-const {Contact} = require('../models/contact');
+const { Contact } = require('../models/contact');
+const { ObjectId } = require('mongodb');
 
 
 const get = async (req, res, next) => {
   const {_id: owner} = req.user;
   const {page = 1, limit = 20} = req.query;
   const skip = (page - 1) * limit;
-  const results = await Contact.find({owner}, '-createdAt -updatedAt', { skip, limit }).populate("owner", "email");
+  const results = await Contact.find({ owner },
+    "-createdAt -updatedAt",
+    {
+      skip,
+      limit,
+    }
+  ).populate("owner", "email");
   
   res.json(results)
 }
@@ -22,8 +29,9 @@ const getById = async (req, res, next) => {
 }
 
 const create = async (req, res) => {
-  const {_id: owner} = req.user;
-  const result = await Contact.create({...req.body, owner});
+  const data = req.body;
+  const owner = req.user._id;
+  const result = await Contact.create({...data, owner });
 
   res.status(201).json(result)
 }
@@ -48,16 +56,15 @@ const updateStatus = async (req, res, next) => {
 }
 
 const remove = async (req, res, next) => {
-  const { id } = req.params
-  const result = await Contact.findByIdAndRemove(id);
+  const { id } = req.params;
+  const owner = req.user._id;
+  const result = await Contact.findByIdAndRemove(id, {owner});
   
   if (!result) {
     throw HttpError(404, 'Not Found')
   }
 
-  res.json({
-    message: 'Delete success'
-  })
+  res.json(result)
 }
 
 module.exports = {
